pokemon-vgc-ai/requirements.txt
# Pokemon VGC AI - Dependencies

# Core ML and RL
poke-env>=0.6.0
stable-baselines3>=2.0
torch>=2.0
gymnasium>=0.29

# Data Processing
pandas>=2.0
numpy>=1.24

# UI Framework
streamlit>=1.28

# Web Scraping
requests>=2.31
beautifulsoup4>=4.12
aiohttp>=3.9.0

# Database
sqlalchemy>=2.0

# Testing
pytest>=7.4
pytest-asyncio>=0.21.0
hypothesis>=6.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
loguru>=0.7.0
rich>=13.0

# Configuration Management (Phase 1)
hydra-core>=1.3
omegaconf>=2.3

# Data Validation (Phase 2)
pydantic>=2.0
pyarrow>=14.0

# Experiment Tracking (Phase 6)
mlflow>=2.9

# Interpretability (Phase 6)
shap>=0.44

# HuggingFace (Data Loading)
huggingface-hub>=0.19.0
datasets>=2.15.0

# Fuzzy String Matching (Pipeline Fixes v2)
rapidfuzz>=3.5

# Action Masking (Pipeline Fixes v2)
sb3-contrib>=2.0
